code cleanup, trying to be more generic
