+++ src/TodoGenieLib/Models/ConfigModel.cs
@@ -2,6 +2,7 @@ namespace TodoGenieLib.Models;
 
 public class ConfigModel {
     public string GithubApiKey { get; set; } = string.Empty;
+    public string GithubEndpoint { get; set; } = string.Empty;
     public string RootDirectory { get; set; } = Environment.CurrentDirectory;
     public string SecretFileName { get; set; } = "secrets.json";
     public string ConfigDirectory { get; set; } = Path.Join(Environment.GetFolderPath(Environment.SpecialFolder.UserProfile),".todogenie");
